+++ src/frontend/terminal.c
@@ -143,7 +143,7 @@ outbufputc(void)
 
     if (ourbuf.count != BUFSIZ) {
       fputs(staticbuf, cp_out);
-      memset(staticbuf, 0, BUFSIZ-ourbuf.count);
+      memset(staticbuf, 0, (size_t) (BUFSIZ - ourbuf.count));
       ourbuf.count = BUFSIZ;
       ourbuf.ptr = staticbuf;
     }
